+++ test/index/constructor.js
@@ -1,0 +1,43 @@
+import chai from 'chai';
+
+import Pivot from '../../src';
+
+chai.expect();
+const expect = chai.expect;
+
+const dataArray = [
+ ['name', 'gender', 'house', 'age'],
+ ['Jon', 'm', 'Stark', 14],
+ ['Arya', 'f', 'Stark', 10],
+ ['Cersei', 'f', 'Baratheon', 38],
+ ['Tywin', 'm', 'Lannister', 67],
+ ['Tyrion', 'm', 'Lannister', 34],
+ ['Joffrey', 'm', 'Baratheon', 18],
+ ['Bran', 'm', 'Stark', 8],
+ ['Jaime', 'm', 'Lannister', 32],
+ ['Sansa', 'f', 'Stark', 12],
+];
+const rowsToPivotTestOne = ['gender', 'name'];
+const colsToPivotTestOne = ['house'];
+const aggregationCategory = 'age';
+const aggregationType = 'sum';
+
+export default () => {
+  it('should create an empty pivot when instantiated without params', () => {
+    const pivot = new Pivot();
+
+    expect(pivot.data).to.deep.equal({});
+  });
+
+  it('should create a pivot when passed data', () => {
+    const pivot = new Pivot(
+      dataArray,
+      rowsToPivotTestOne,
+      colsToPivotTestOne,
+      aggregationCategory,
+      aggregationType,
+    );
+
+    expect(pivot.data.table).to.have.length.above(0);
+  });
+};
