+++ test/index/toggle.js
@@ -1,0 +1,41 @@
+import chai from 'chai';
+
+import Pivot from '../../src';
+
+chai.expect();
+const expect = chai.expect;
+
+const dataArray = [
+ ['name', 'gender', 'house', 'age'],
+ ['Jon', 'm', 'Stark', 14],
+ ['Arya', 'f', 'Stark', 10],
+ ['Cersei', 'f', 'Baratheon', 38],
+ ['Tywin', 'm', 'Lannister', 67],
+ ['Tyrion', 'm', 'Lannister', 34],
+ ['Joffrey', 'm', 'Baratheon', 18],
+ ['Bran', 'm', 'Stark', 8],
+ ['Jaime', 'm', 'Lannister', 32],
+ ['Sansa', 'f', 'Stark', 12],
+];
+const rowsToPivotTestOne = ['gender', 'name'];
+const colsToPivotTestOne = ['house'];
+const aggregationCategory = 'age';
+const aggregationType = 'sum';
+
+export default () => {
+  it('should return table to normal state when toggled twice', () => {
+    const pivot = new Pivot(
+      dataArray,
+      rowsToPivotTestOne,
+      colsToPivotTestOne,
+      aggregationCategory,
+      aggregationType,
+    );
+
+    const startState = pivot.data.table;
+
+    pivot.toggle(1).toggle(1);
+
+    expect(startState).to.deep.equal(pivot.data.table);
+  });
+};
